+++ oidc_provider/tests/app/utils.py
@@ -44,7 +44,7 @@ def create_fake_user():
     return user
 
 
-def create_fake_client(response_type, is_public=False, require_consent=True):
+def create_fake_client(response_type, is_public=False, require_consent=True, require_pkce=False):
     """
     Create a test client, response_type argument MUST be:
     'code', 'id_token' or 'id_token token'.
@@ -62,6 +62,7 @@ def create_fake_client(response_type, is_public=False, require_consent=True):
     client.redirect_uris = ['http://example.com/']
     client.require_consent = require_consent
     client.scope = ['openid', 'email']
+    client.require_pkce = require_pkce
     client.save()
 
     # check if response_type is a string in a python 2 and 3 compatible way